requirements.txt
# Core libraries for PV simulation and data handling
pvlib==0.11.1
pandas==2.2.2
numpy==1.26.4

# For data validation and CLI
pydantic==2.8.2
typer==0.12.3
PyYAML==6.0.1

# For data visualization
matplotlib==3.9.1

# For API requests
requests==2.32.3

# For testing
pytest==8.3.2